backend/requirements.txt
fastapi==0.135.2
uvicorn[standard]==0.42.0
torch
torchvision
timm==1.0.26
pillow==12.1.1
numpy==2.4.4
python-multipart==0.0.22
slowapi==0.1.9
